refactor: moved the window and openfin closing to action creator
