+++ test/child/reducers/windowStateSpec.js
@@ -1,12 +1,6 @@
 import { expect } from 'chai';
-import sinon from 'sinon';
 import windowState from '../../../src/child/reducers/windowState.js';
 import { WINDOW as ACTION_TYPES } from '../../../src/child/constants/actionTypes';
-
-const publish = sinon.spy();
-const close = sinon.spy();
-global.fin = { desktop: { InterApplicationBus: { publish } } };
-global.window = { close };
 
 describe('child/reducers/windowState', () => {
     it('should return the initial state', () => {
@@ -84,12 +78,6 @@ describe('child/reducers/windowState', () => {
             isCompact: false,
             isMaximised: false
         });
-
-        expect(publish.calledOnce).to.be.true;
-        expect(close.calledOnce).to.be.true;
-        expect(publish.calledWithExactly('childClosing', { id: window.id })).to.be.true;
-        expect(close.calledWithExactly()).to.be.true;
-
         expect(windowState({
             isCompact: true,
             isMaximised: false
@@ -97,12 +85,6 @@ describe('child/reducers/windowState', () => {
             isCompact: true,
             isMaximised: false
         });
-
-        expect(publish.calledTwice).to.be.true;
-        expect(close.calledTwice).to.be.true;
-        expect(publish.calledWithExactly('childClosing', { id: window.id })).to.be.true;
-        expect(close.calledWithExactly()).to.be.true;
-
         expect(windowState({
             isCompact: false,
             isMaximised: true
@@ -110,12 +92,6 @@ describe('child/reducers/windowState', () => {
             isCompact: false,
             isMaximised: true
         });
-
-        expect(publish.calledThrice).to.be.true;
-        expect(close.calledThrice).to.be.true;
-        expect(publish.calledWithExactly('childClosing', { id: window.id })).to.be.true;
-        expect(close.calledWithExactly()).to.be.true;
-
         expect(windowState({
             isCompact: true,
             isMaximised: true
@@ -123,11 +99,6 @@ describe('child/reducers/windowState', () => {
             isCompact: true,
             isMaximised: true
         });
-
-        expect(publish.callCount).to.equal(4);
-        expect(close.callCount).to.equal(4);
-        expect(publish.calledWithExactly('childClosing', { id: window.id })).to.be.true;
-        expect(close.calledWithExactly()).to.be.true;
     });
 
     it('should return the previous state for any unknown action', () => {
